Remove loginhelpers class and move functionality into login page
